requirements.txt
requests
flask
python-dotenv
gunicorn
flask_cors
boto3
pyairtable